refactor proof pack ai evidence router
