+++ src/app/router_registry.py
@@ -164,6 +164,9 @@ from app.routers.dpm_command_center_portfolio_memory_search import (
 from app.routers.dpm_construction import router as dpm_construction_router
 from app.routers.dpm_construction_actions import router as dpm_construction_actions_router
 from app.routers.dpm_proof_pack_ai import router as dpm_proof_pack_ai_router
+from app.routers.dpm_proof_pack_ai_evidence_input import (
+    router as dpm_proof_pack_ai_evidence_input_router,
+)
 from app.routers.dpm_proof_pack_evidence import router as dpm_proof_pack_evidence_router
 from app.routers.dpm_proof_packs import router as dpm_proof_packs_router
 from app.routers.dpm_wave_actions import router as dpm_wave_actions_router
@@ -493,6 +496,7 @@ def register_routers(app: FastAPI) -> None:
     app.include_router(dpm_construction_actions_router)
     app.include_router(dpm_proof_packs_router)
     app.include_router(dpm_proof_pack_evidence_router)
+    app.include_router(dpm_proof_pack_ai_evidence_input_router)
     app.include_router(dpm_proof_pack_ai_router)
     app.include_router(dpm_waves_router)
     app.include_router(dpm_wave_lookup_router)
